If headline text is empty, allow skipping headline
This commit makes the string comparison avoid failure when the title component
is nil, (it's not an empty string).

This occurs for example when a headline is just a TODO cookie with no later
headline text.

The two blocks for generating the link properly attempt to skip headlines where
there is no 4th title component, so later generation correctly handles the case
where an empty link generation would be attempted in the TOC.

Note, empty links would not render or behave, so forcing their generation would
be incorrect anyway.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>